restruct(uutils): clean user utilities implementation directory
The old files contained in this repository are not used, or have suboptimal
implementations for some things. I have cleaned them up here and adjusted the
name of the `edit` user-utilities module to `text`, as this is more descriptive
of what the module is supposed to do.
