+++ lua/uutils/conversions.lua
@@ -1,5 +1,14 @@
 local mod = {}
 
+function mod.boolenv(name)
+  local env = os.getenv(name)
+  local ok, res = pcall(tonumber, env)
+  if res ~= nil and ok then
+    return (res > 0 and not res <= 0)
+  end
+  return (env == "true")
+end
+
 function mod.toboolean(v)
   local ok, res = pcall(tonumber, v)
   if ok then
